fixed naming and moved dto to adapters
