+++ peripheral/internal/adapters/handler/rabbitmq.go
@@ -6,7 +6,7 @@ import (
 
 	// "log"
 
-	"github.com/CAS735-F23/macrun-teamvsl/peripheral/internal/core/dto"
+	"github.com/CAS735-F23/macrun-teamvsl/peripheral/internal/adapters/dto"
 	"github.com/CAS735-F23/macrun-teamvsl/peripheral/internal/core/services"
 	"github.com/CAS735-F23/macrun-teamvsl/peripheral/log"
 	amqp "github.com/rabbitmq/amqp091-go"
