+++ peripheral/internal/core/services/services.go
@@ -3,8 +3,8 @@ package services
 import (
 	"log"
 
+	"github.com/CAS735-F23/macrun-teamvsl/peripheral/internal/adapters/dto"
 	"github.com/CAS735-F23/macrun-teamvsl/peripheral/internal/core/domain"
-	"github.com/CAS735-F23/macrun-teamvsl/peripheral/internal/core/dto"
 	"github.com/CAS735-F23/macrun-teamvsl/peripheral/internal/core/ports"
 	"github.com/google/uuid"
 	"go.uber.org/zap"
